factor out store and reterieve passwords into util
